# shared/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoSense Shared Contract",
    "description": "GeoSense AI — Internal AI Engine API (Python FastAPI)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/risks/jobs": {
      "post": {
        "tags": [
          "risks"
        ],
        "summary": "Submit Risk Job",
        "description": "Submit a risk assessment job (§3.3.1).",
        "operationId": "submit_risk_job_api_v1_risks_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_JobAccepted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/jobs/{pythonJobId}": {
      "get": {
        "tags": [
          "risks"
        ],
        "summary": "Get Risk Status",
        "description": "Poll risk job status (§3.3.2).",
        "operationId": "get_risk_status_api_v1_risks_jobs__pythonJobId__get",
        "parameters": [
          {
            "name": "pythonJobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pythonjobid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RiskJobData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risks/jobs": {
      "post": {
        "tags": [
          "client: risks"
        ],
        "summary": "Submit Risk Job",
        "description": "Submit a risk assessment job (§2.5.1).",
        "operationId": "submit_risk_job_api_risks_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskJobSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_JobQueued_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoundingBox": {
        "properties": {
          "minX": {
            "type": "number",
            "title": "Minx"
          },
          "minY": {
            "type": "number",
            "title": "Miny"
          },
          "maxX": {
            "type": "number",
            "title": "Maxx"
          },
          "maxY": {
            "type": "number",
            "title": "Maxy"
          }
        },
        "type": "object",
        "required": [
          "minX",
          "minY",
          "maxX",
          "maxY"
        ],
        "title": "BoundingBox",
        "description": "WGS84 bounding box (Appendix B.1)."
      },
...
      "Envelope_JobAccepted_": {
        "properties": {
          "statusCode": {
            "type": "integer",
            "title": "Statuscode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobAccepted"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "title": "Envelope[JobAccepted]"
      },
      "Envelope_JobQueued_": {
        "properties": {
          "statusCode": {
            "type": "integer",
            "title": "Statuscode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobQueued"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "title": "Envelope[JobQueued]"
      },
...
      "Envelope_RiskJobData_": {
        "properties": {
          "statusCode": {
            "type": "integer",
            "title": "Statuscode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskJobData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "title": "Envelope[RiskJobData]"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Individual error object within the errors array (§1.5)."
      },
...
      "GeoJsonPolygon": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Polygon"
          },
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "GeoJsonPolygon",
        "description": "GeoJSON Polygon geometry (Appendix B.2)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobAccepted": {
        "properties": {
          "pythonJobId": {
            "type": "string",
            "title": "Pythonjobid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "acceptedAt": {
            "type": "string",
            "title": "Acceptedat"
          }
        },
        "type": "object",
        "required": [
          "pythonJobId",
          "acceptedAt"
        ],
        "title": "JobAccepted",
        "description": "202 Accepted payload for internal AI-engine job submission (§3.x.1)."
      },
      "JobQueued": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "parcelId": {
            "type": "string",
            "title": "Parcelid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "estimatedDuration": {
            "type": "string",
            "title": "Estimatedduration",
            "default": "2-6 hours"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "parcelId"
        ],
        "title": "JobQueued",
        "description": "Async job accepted payload (§1.3 / all §2 POST /jobs)."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
...
      "RiskJobData": {
        "properties": {
          "pythonJobId": {
            "type": "string",
            "title": "Pythonjobid"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "results": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskResults"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "pythonJobId",
          "status"
        ],
        "title": "RiskJobData"
      },
      "RiskJobRequest": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "parcelId": {
            "type": "string",
            "title": "Parcelid"
          },
          "geoJson": {
            "$ref": "#/components/schemas/GeoJsonPolygon"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "riskTypes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risktypes",
            "default": [
              "flood",
              "seismic",
              "expansiveSoil",
              "liquefaction"
            ]
          },
          "soilData": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SoilData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "parcelId",
          "geoJson"
        ],
        "title": "RiskJobRequest"
      },
      "RiskJobSubmit": {
        "properties": {
          "parcelId": {
            "type": "string",
            "title": "Parcelid"
          },
          "riskTypes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risktypes",
            "default": [
              "flood",
              "seismic",
              "expansiveSoil",
              "liquefaction"
            ]
          }
        },
        "type": "object",
        "required": [
          "parcelId"
        ],
        "title": "RiskJobSubmit"
      },
      "RiskResults": {
        "properties": {
          "overallScore": {
            "type": "integer",
            "title": "Overallscore"
          },
          "flood": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTypeResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "seismic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTypeResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "expansiveSoil": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTypeResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "liquefaction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTypeResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "overallScore"
        ],
        "title": "RiskResults"
      },
      "RiskTypeResult": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "zonesGeoJson": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zonesgeojson"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "type": "object",
        "required": [
          "score",
          "level"
        ],
        "title": "RiskTypeResult"
      },
...
      "SoilData": {
        "properties": {
          "clayContent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claycontent"
          },
          "sandContent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandcontent"
          },
          "waterTableDepth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watertabledepth"
          }
        },
        "type": "object",
        "title": "SoilData",
        "description": "Soil data passed to risk assessment."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}